refactor(draft): decompose DraftPage into helpers, hooks, and sub-components + add undo/redo keyboard shortcuts
